Add net to screen, harmonize colors, scale sizes
The net is a rect drawn to the screen. The net's thickness is 4% of the
screen width. This is to allow the game to scale to varios resolutions
easily.

The paddle sizes now also scale at a default of 3% thick, 13% tall.

a new self.equipment color harmonizes the color across the paddle net,
and play/pause button text. This is in order to stay more true to the
orignal game's color scheme of white on black. One exception is the ball
color, which is clightly darker.