Fine tune some sprite placements
Paddles when reaching the top or bottom of the screen now have a buffer
of 5 pixels. This visually improves the placement of the paddles.
Previously, when reaching the edge of the screen, the paddles seemed to
go off screen.

Improved placement of the paddles from the edge of the screen.
Previously, we used an arbatrary  magic constant of 20. Now, paddles
will be a paddle's width from the edge of the screen.

Visually improved how computer player (Tc) follows the ball. Now, the
paddles centery follows the ball's same, rather than using simply the y
axis on its own. This was a small oversight of mine.