Remove player_misses feature
This new branch 'lives-only' is a slightly simplified version of the
game. I have removed the feature of allowing n misses of the ball before
losing a life.

If I kept that feature, I'd have to display both misses and lives
remaining on screen, in addition to the score, and, in the future, the
current level number.