Display score and lives on screen
Score and lives are now displayed on screen and on the correct side of
the screen for each player.

This is the first really playable version of the game. I just need to
add instructions, 'press q to quit, p to pause', sound effects. Then I
will have a fully ready beta version.