(0.0.12) Paddle speed = relative to frame_rate
This is not the best method, probably. I should have the paddle's and ball's
speed independent of the frame rate.

But, that's all part of the fun in the journey of programming. Try
things out, make one small change and test it.

KEEPING_AT_IT = True
IT_DOESNT_WORK = False

try: # out new things
    while KEEPING_AT_IT:
        print('Make a small change and test it...')
        if IT_DOESNT_WORK:
            print('You still learned something. Keep at it!')
except KeyboardInterrupt:
    print('See you again soon!')